Modularize latexindent and remove warning-only os check
